style(readme): remove underline from images
This commit updates the `README.md` file to remove the underline from
the images that are wrapped in anchor tags. The underline was caused by
the default styling of the anchor tags and was making the images look
less appealing.

The underline was removed by adding an inline style
`text-decoration:none;` to the anchor tags. This overrides the default
styling and removes the underline.